+++ test/libsolidity/smtCheckerTests/array_members/length_1d_copy_2d_storage_to_memory.sol
@@ -1,0 +1,11 @@
+pragma experimental SMTChecker;
+pragma experimental ABIEncoderV2;
+
+contract C {
+	uint[][] arr;
+	function f() public view {
+		uint[][] memory arr2 = arr;
+		assert(arr2[0].length == arr[0].length);
+		assert(arr2.length == arr.length);
+	}
+}