+++ test/libsolidity/smtCheckerTests/loops/for_loop_array_assignment_memory_memory.sol
@@ -1,11 +1,12 @@
 pragma experimental SMTChecker;
 
 contract LoopFor2 {
-	function testUnboundedForLoop(uint n, uint[] memory b, uint[] memory c) public pure {
+	function testUnboundedForLoop(uint[] memory b, uint[] memory c) public pure {
 		b[0] = 900;
 		uint[] memory a = b;
-		require(n > 0 && n < 100);
-		for (uint i = 0; i < n; i += 1) {
+		require(b.length == c.length);
+		require(b.length > 0 && b.length < 100);
+		for (uint i = 0; i < b.length; i++) {
 			b[i] = i + 1;
 			c[i] = b[i];
 		}
@@ -15,6 +16,7 @@ contract LoopFor2 {
 	}
 }
 // ----
-// Warning: (281-301): Assertion violation happens here
-// Warning: (305-324): Assertion violation happens here
-// Warning: (328-347): Assertion violation happens here
+// Warning: (295-300): Overflow (resulting value larger than 2**256 - 1) happens here
+// Warning: (324-344): Assertion violation happens here
+// Warning: (348-367): Assertion violation happens here
+// Warning: (371-390): Assertion violation happens here