+++ test/libsolidity/smtCheckerTests/loops/while_loop_array_assignment_memory_memory.sol
@@ -1,12 +1,13 @@
 pragma experimental SMTChecker;
 
 contract LoopFor2 {
-	function testUnboundedForLoop(uint n, uint[] memory b, uint[] memory c) public pure {
+	function testUnboundedForLoop(uint[] memory b, uint[] memory c) public pure {
 		b[0] = 900;
 		uint[] memory a = b;
-		require(n > 0 && n < 100);
+		require(b.length == c.length);
+		require(b.length > 0 && b.length < 100);
 		uint i;
-		while (i < n) {
+		while (i < b.length) {
 			b[i] = i + 1;
 			c[i] = b[i];
 			++i;
@@ -16,9 +17,9 @@ contract LoopFor2 {
 		assert(b[0] == 900);
 	}
 }
-// ====
-// SMTSolvers: z3
 // ----
-// Warning: (281-301): Assertion violation happens here
-// Warning: (305-324): Assertion violation happens here
-// Warning: (328-347): Assertion violation happens here
+// Warning: (290-295): Overflow (resulting value larger than 2**256 - 1) happens here
+// Warning: (316-319): Overflow (resulting value larger than 2**256 - 1) happens here
+// Warning: (327-347): Assertion violation happens here
+// Warning: (351-370): Assertion violation happens here
+// Warning: (374-393): Assertion violation happens here